Pulled DateCube out as a component to reuse + some styling fot the FeaturedPostCard
