+++ src/ui/components/DateCube/DateCube.module.scss
@@ -1,0 +1,49 @@
+@import "src/utility/responsive";
+@import "src/utility/variables";
+@import "src/utility/functions";
+
+$dateOffset: 20px;
+
+.date {
+  align-items: center;
+  color: $white;
+  display: flex;
+  flex-direction: column;
+  height: 60px;
+  justify-content: center;
+  padding: 16px;
+  position: absolute;
+  left: 18px;
+  letter-spacing: 0.8px;
+  line-height: 1.15;
+  text-align: center;
+  bottom: 100%;
+  transform: translateY(calc(100% - #{$dateOffset}));
+  width: 64px;
+
+  .root.turquoise & {
+    background: $turquoise;
+  }
+
+  .root.yellow & {
+    background: $pastelYellow;
+  }
+
+  .root.pink & {
+    background: $pastelPink;
+  }
+
+  .root.purple & {
+    background: $pastelPurple;
+  }
+}
+
+.dateNumber {
+  font-family: $serif;
+  font-weight: bold;
+  font-size: 20px;
+}
+
+.dateString {
+  font-size: 14px;
+}
