+++ src/ui/components/EventCard/EventCard.module.scss
@@ -14,14 +14,14 @@ $dateOffset: 20px;
   position: relative;
   padding: between(26px, 40px, $sm, $xlg) between(20px, 30px, $sm, $xlg);
   height: 512px;
-  
+
   @include respond-to(onlySmall) {
     height: 400px;
     padding: 40px 30px;
   }
-  
+
   &::after {
-    content: '';
+    content: "";
     position: absolute;
     bottom: 0;
     left: 0;
@@ -64,50 +64,6 @@ $dateOffset: 20px;
   .root.purple & {
     color: $pastelPurple;
   }
-}
-
-.date {
-  align-items: center;
-  color: $white;
-  display: flex;
-  flex-direction: column;
-  height: 60px;
-  justify-content: center;
-  padding: 16px;
-  position: absolute;
-  left: 18px;
-  letter-spacing: 0.8px;
-  line-height: 1.15;
-  text-align: center;
-  bottom: 100%;
-  transform: translateY(calc(100% - #{$dateOffset}));
-  width: 64px;
-
-  .root.turquoise & {
-    background: $turquoise;
-  }
-
-  .root.yellow & {
-    background: $pastelYellow;
-  }
-
-  .root.pink & {
-    background: $pastelPink;
-  }
-
-  .root.purple & {
-    background: $pastelPurple;
-  }
-}
-
-.dateNumber {
-  font-family: $serif;
-  font-weight: bold;
-  font-size: 20px;
-}
-
-.dateString {
-  font-size: 14px;  
 }
 
 .title {
