+++ src/ui/components/FeaturedPostCard/FeaturedPostCard.js
@@ -1,9 +1,10 @@
 import React from "react";
 import styles from "./FeaturedPostCard.module.scss";
+import DateCube from "src/ui/components/DateCube";
 import { H2, Body3 } from "../Typography";
 import Arrow from "../Arrow";
 
-const FeaturedPostCard = ({ title, text, imageUrl, url, fullWidth }) => (
+const FeaturedPostCard = ({ date, title, text, imageUrl, url, fullWidth }) => (
   <div className={`${styles.root} ${fullWidth ? styles.fullWidth : ""}`}>
     <div
       className={styles.image}
@@ -15,6 +16,9 @@ const FeaturedPostCard = ({ title, text, imageUrl, url, fullWidth }) => (
       rel="noopener noreferrer"
       className={styles.card}
     >
+      {/* TODO: replace hardcoded values */}
+      {console.log(date)}
+      <DateCube date={"13 may"} color={"red"} />
       <H2 bottom="small">{title}</H2>
       <Body3 light>{text}</Body3>
       <Arrow className={styles.arrow} top="medium" />
